Move mapViewModel
